+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/InvoicePanel.java
@@ -31,7 +31,7 @@ public final class InvoicePanel extends EntityPanel {
                     .includeDetailTabPane(false)
                     .includeDetailControls(false)
                     .build());
-    tablePanel().editableAttributes().remove(Invoice.TOTAL);
+    tablePanel().configure().editableAttributes().remove(Invoice.TOTAL);
     addDetailPanel(invoiceLinePanel);
   }
 }