+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityDependenciesPanel.java
@@ -124,7 +124,7 @@ public final class EntityDependenciesPanel extends JPanel {
         return popupMenuControls;
       }
     };
-    tablePanel.setIncludeConditionPanel(false);
+    tablePanel.configure().includeConditionPanel(false);
 
     return tablePanel.initialize();
   }